requirements.txt
discord.py==2.0.1
requests==2.27.1
